Reorganize MessageInvokerPoolTests
