+++ test/Ocelot.UnitTests/SequentialTests.cs
@@ -1,6 +1,4 @@
-﻿using System.Runtime.InteropServices;
-
-namespace Ocelot.UnitTests;
+﻿namespace Ocelot.UnitTests;
 
 /// <summary>
 /// Apply <see cref="CollectionAttribute"/> to classes to disable parallelization.
@@ -8,21 +6,6 @@ namespace Ocelot.UnitTests;
 [CollectionDefinition(nameof(SequentialTests), DisableParallelization = true)]
 public class SequentialTests
 {
-    /// <summary>
-    /// Unstable <see cref="Requester.MessageInvokerPoolTests"/>.
-    /// </summary>
-    [Collection(nameof(SequentialTests))]
-    public class MessageInvokerPoolTests
-    {
-        [Fact]
-        public async Task Should_reuse_cookies_from_container()
-        {
-            // Test #1
-            var test = new Requester.MessageInvokerPoolTests();
-            await test.Should_reuse_cookies_from_container();
-        }
-    }
-
     /// <summary>
     /// Unstable <see cref="Consul.ConsulTests"/>.
     /// </summary>
